deinline acb IO
