+++ src/acb.h
@@ -981,49 +981,17 @@ _acb_vec_scalar_div_fmpz(acb_ptr res, acb_srcptr vec, slong len, const fmpz_t c,
         acb_div_fmpz(res + i, vec + i, c, prec);
 }
 
-ACB_INLINE void
-acb_fprint(FILE * file, const acb_t x)
-{
-    flint_fprintf(file, "(");
-    arb_fprint(file, acb_realref(x));
-    flint_fprintf(file, ", ");
-    arb_fprint(file, acb_imagref(x));
-    flint_fprintf(file, ")");
-}
-
-ACB_INLINE void
-acb_print(const acb_t x)
-{
-    acb_fprint(stdout, x);
-}
-
+#ifdef FLINT_HAVE_FILE
+void acb_fprint(FILE * file, const acb_t x);
 void acb_fprintd(FILE * file, const acb_t z, slong digits);
-
-ACB_INLINE void
-acb_printd(const acb_t z, slong digits)
-{
-    acb_fprintd(stdout, z, digits);
-}
-
 void acb_fprintn(FILE * fp, const acb_t z, slong digits, ulong flags);
+#endif
 
-ACB_INLINE void
-acb_printn(const acb_t x, slong digits, ulong flags)
-{
-    acb_fprintn(stdout, x, digits, flags);
-}
+void acb_print(const acb_t x);
+void acb_printd(const acb_t z, slong digits);
+void acb_printn(const acb_t x, slong digits, ulong flags);
 
-ACB_INLINE void
-_acb_vec_printn(acb_srcptr vec, slong len, slong ndigits, ulong flags)
-{
-    slong i;
-    for (i = 0; i < len; i++)
-    {
-        acb_printn(vec + i, ndigits, flags);
-        if (i < len - 1)
-            flint_printf(", ");
-    }
-}
+void _acb_vec_printn(acb_srcptr vec, slong len, slong ndigits, ulong flags);
 
 void acb_randtest(acb_t z, flint_rand_t state, slong prec, slong mag_bits);
